+++ relay-server/src/processing/logs/process.rs
@@ -45,6 +45,10 @@ pub fn expand(logs: Managed<SerializedLogs>, _ctx: Context<'_>) -> Managed<Expan
             headers: logs.headers,
             #[cfg(feature = "processing")]
             retention: _ctx.project_info.config.event_retention,
+            #[cfg(feature = "processing")]
+            serialize_meta_attrs: _ctx
+                .project_info
+                .has_feature(relay_dynamic_config::Feature::OurLogsMetaAttributes),
             logs: all_logs,
         }
     })