+++ relay-server/src/services/projects/cache/project.rs
@@ -78,13 +78,10 @@ impl<'a> Project<'a> {
         let current_limits = self.rate_limits().current_limits();
 
         let quotas = state.as_deref().map(|s| s.get_quotas()).unwrap_or(&[]);
-        let current_limits_clone = current_limits.clone();
-        let envelope_limiter =
-            EnvelopeLimiter::new(CheckLimits::NonIndexed, move |item_scoping, _| {
-                let current_limits_clone = current_limits_clone.clone();
-
-                async move { Ok(current_limits_clone.check_with_quotas(quotas, item_scoping)) }
-            });
+        let envelope_limiter = EnvelopeLimiter::new(CheckLimits::NonIndexed, |item_scoping, _| {
+            let current_limits = Arc::clone(&current_limits);
+            async move { Ok(current_limits.check_with_quotas(quotas, item_scoping)) }
+        });
 
         let (mut enforcement, mut rate_limits) = envelope_limiter
             .compute(envelope.envelope_mut(), &scoping)